Upgrade address rotation to a real shuffle
Historically, it was common for DNS servers to do a round robin
rotation of address records as they're served, to act as a poor
sort of load distribution aid by seeding different caches and
users with different entries at the top.

It's historically common for some software to either only use the
first address, or to start with the first and then proceed
linearly for failover.  In these linear failover cases, however,
the distribution of failed-over requests is very uneven, and this
can be fixed by replacing rotation with true shuffling, which is
what we're doing with this commit.

The longer explanation:

It's arguable that for an auth-only server, all of this shouldn't
matter in an ideal world, since all of our clients are other DNS
servers (recursors) which should also be handling this as they
hand out results to their clients.  However, it's probably not
wise to rely on them doing so in a perfectly unbiased way (or at
all), so we should do it defensively in the authserver as well.

If we're going to handle it in the authserver at all, then the
rotation strategy isn't a very good one in the face of linear
client failover due to address failure, as it tends to distribute
the fallback load very unevenly.

Consider an example:

foo A 192.0.2.1
foo A 192.0.2.2
foo A 192.0.2.3

If we merely rotate these outputs, there are 3 possible orderings
we can serve: 123, 231, and 312, which gives the best possible
distribution of first options.

However, if the first address fails, 100% of the linear-failover
clients which received the 123 result will fail over to the second
address, because with all of the 1's removed, the three options
effectively become: 23, 23, 32.  This results in the undesirable
outcome that 2/3 of clients connect to the second address while
1/3 connect to the third.

If, instead of rotating, we actually randomly shuffle our outputs,
we will serve 6 possible orderings: 123, 132, 213, 231, 312, 321.
Following the same example as above, when the first address fails,
the affected clients would fail over as an even split between the
second and third choices, thus keeping the load balance as optimal
as such a scheme is capable of.

It also turns out to be true that even if client caches end up
implementing the simpler rotation strategy at their own level, on
top of our shuffling, the initial shuffle still prevents their
rotations from re-creating the imbalance, at least in the global
aggregate of all such caches.

--

While we're implementing this, we've also split the runtime
shuffling of the records from their encoding, as this fits better
with future plans to move the encoding part offline from the
runtime query flow, and it keeps the code simpler anyways.