Pre-encode response RR data
The old RR data processing model, prior to this commit, was that
the ltree data structure had separate rrset and rdata types for
all of the major RR types supported, and the network response
code, after parsing a live question from the network, was
responsible for encoding that data into the wire response format
via a set of per-RR-type encoding functions while handling
compression (and in the case of NS, glue attachment) in realtime.

The new model created by this commit collapses most (other than
DYNA and DYNC) of the ltree RR data types into a single "raw" type
and pre-encodes their wire format response data at zonefile
parsing time, including compression and glue.  The runtime network
response code now just does a memcpy() of a pre-encoded block of
wire response data that is already compressed and glued as
appropriate, and then in some cases (delegation responses,
negative SOA responses, and the RHS of wildcard CNAME and MX)
processes an array of compression offset fixups, where a small
integer has to be added to a set of existing compression pointers
in the copied response data to make up for the length of the
actual query name (which is indeterminate in these cases at load
time).

Some of the key impact highlights:
* Fewer different data structures within ltree
* Data structures are simpler in general
* The work of wire-format encoding, NS glue lookups, and DNS
  compression is moved out of the realtime response-generation
  path, and is instead done offline once per rrset at load time

Other visible things that changed while working on this:
* Our data output size limit of ~16K for responses is now checked
  accurately against the post-compression data sizes, whereas
  before it was checked heuristically and did not account for
  right-hand-side compression savings, and so tended to reject
  some edge cases pointlessly
* NS rrsets are no longer artificially limited to 64 RRs, which
  was a constraint imposed by the old realtime gluing system
* All rrsets are now limited to 1024 RRs per rrset in the general
  case.  Technically it's very hard to reach 1024 RRs without
  exceeding our ~16KB response size limit anyways, but this makes
  it simpler to trim some bad cases early in the parsing process
* Compression is more aggressive than it was before at finding
  matches, because we can afford it when it's done offline only
  once at load time.
* Some zone data warnings we previously issued about bad smells or
  practices which aren't actually protocol violations have been
  dropped, including: PTR to same zone, MX/SRV to same zone with
  no target address, and SOA master != any zone NS.
* Right hand side names in PTR RRs are once again subject to
  compression against qname and themselves, although it's not
  commonly useful
* ltree limits on the number of child nodes for a domainname are
  actually enforced correctly now (fatal for the zone, but not for
  the daemon), in the rare (but possible!) case someone tries to
  add billions of names at one level.

Other internal changes:
* The "label" field of ltree nodes is now the "dname" field.
  While it can be used to read the label of the node as before,
  and this is what happens during tree lookups for the hash table,
  it actually contains the whole domainname.
* The "ltarena" system for allocating per-zone space for small
  label objects has been eliminated, and the "name" data above is
  now just malloc'd the standard way.  ltarena was already making
  less and less sense lately, but the changes in this commit
  really made it seem like a net loss in most scenarios.
* Some functions that used the "lstack" form of domainname data
  now use either dname or raw wire names, making things a little
  simpler.
* zone_t is gone as a type (not needed anymore)
* Everything (nodes and data) beneath a delegation node is deleted
  from the tree, except for the singular NS rrset for the
  delegation itself, after all glue has been consumed and copied
  by such NS rrsets.

Code quality improvements:
* Various #define schemes for templating per-rr-type function
  definitions eliminated in ltree.c and dnspacket.c
* Function pointers and lookup tables of function pointers in both
  files eliminated.
* Net reduction of ~700 SLOC overall, the vast majority of it
  being a reduction in the critical runtime code in dnspacket.c,
  which lost ~1/3 of its total SLOC.