Bugfix for compression case-insensitivity
This bug was causing inflated packet sizes for some responses when
the client-provided query name had uppercase ASCII characters.
The compression code was failing to make use of the lowercase
normalized copy we had available for this purpose.

The impact tends to minor, as this only affects compressible
right-hand-side names in MX, NS, SOA, and CNAME responses (and
never the left-hand-side repeat of the query name for any RR
type), and after the first compression miss from this bug, we
implicitly acquire a pointer to a normalized copy from the zone
data to use from there forward.

In the (now updated) compression torture regression test, it
inflated the total packet size from 569 bytes to 588 bytes, and in
general the impact is limited to at most an inflation by the whole
size of the query name.  Output packet size sanity constraints
never contemplated the savings from this kind of compression
anyways, so there's no deeper fault that can be triggered, either.

(cherry picked from commit a9f62c2de97a2c4cf496857371b6058e3af80acd)