Deprecate DYNC returning A/AAAA
This commit just adds LOG_WARN -level message outputs whenever a
daemon's configuration allows DYNC to return address records.
Support for these configurations will be removed in 4.x.

This was discussed before in an obscure corner of the
documentation.  In the long term view, the current "plugins" are
all doomed and will be replaced by a better design in future
major-version updates, but there will be some time period during
which the old and new designs overlap for user transition.

In the medium-term view, this deprecation allows us to eliminate
DYNC->A/AAAA in the 4.x prototype branches, which allows some
other work there to proceed in a more straightforward manner with
less legacy constraint.