Remove packed attribute from wire_dns_header_t
This has historically had the packed attribute because we were
commonly casting raw buffers to this type with no guarantees about
the buffer alignment.  However, we now do have buffers that are
aligned correctly, and thus commit creates a new union data type
which makes this clear to the compiler and analyzers, thus
allowing us to safely drop assumptions about lack of alignment.