Bugfix for TTL clamping under zones_strict_data
The config settings "min_ttl" and "max_ttl" are intended to set an
allowed range of TTL values, defaulting to 5 and 3600000 (~42d).
TTL values outside of this range are clamped to the range during
zonefile loading with a logged warning message.  Also, the config
setting "zones_strict_data" (which can also be indirectly set by
the CLI flag "-S") is intended to upgrade all zone data warnings
(including the ones above about TTL clamping) to full-on errors,
causing the zonefile to fail to load.

So long as "zones_strict_data" is false, this was all working as
intended.  However, if "zones_strict_data" were true and an out of
range TTL is detected, the bug being fixed in this commit would
both (a) fail to upgrade the TTL clamp warning to a full-on error
and (b) also set the out-of-range TTL to the value "1" instead of
clamping it to the range as intended.

This commit fixes (b) and leaves (a) as the newly-intended
behavior: that the warning about having to clamp a TTL value does
not cause failure under zones_strict_data (which is reasonable
given the data gets fixed by the clamping anyways, unlike other
such cases where the data cannot be fixed by the code).  Going the
other way ("fixing" (a) as well) would have been more invasive and
risky at this point.

This bug was introduced in v3.2.0 in commit 4818a260