RRSet sorting and de-duplication
It's always been true (but not hugely important) that we should
have been eliminating duplicate RRs (identical rdata at same name,
class, and type).  DNSSEC absolutely requires it, and further it
requires that we're able to sort the RRs of an RRSet in canonical
order for hashing.

This commit moves more of the rdata encoding work from the ltree
code to the parser code, and restructures things such that we
initially create an array of rdata per rrset as we parse the
zonefile and then compile the pre-encoded response data from those
arrays during ltree post-processing.

Having this intermediate array-of-rdata form causes some extra
work, but it gives us the opportunity to sort and deduplicate the
rrsets of the zone efficiently.  The rdata items are sorted on
insertion into the array from the parser, and duplicates are
detected at insertion time as well.  Dupes cause a warning about
ignoring the duplicate data, or a hard error if zones_strict_data
is true.

Either way, no duplicates are actually allowed in the resulting
rrset, and the rrset ends up always being stored in the ltree in
DNSSEC-sorted order.

While some previous work has been slowly moving things in the
right directions to get here, this is the first commit to gdnsd
that really goes out of its way to enable future DNSSEC efforts to
happen.  Baby steps!