node and zone type refactoring
This brings back the shadow of the old zone_t, but replaces it
with the subclassing of ltree_node_t as ltree_node_zroot_t, which
is set up in a similar fashion to how we subclass ltree_rrset_t.

This allows us to add extra zone metadata without inflating all
the rest of the nodes.  For now this is just used to hold the
serial value from the SOA, but DNSSEC will also want to make use
of zone metadata.

While we're here, also split the zone_cut flag into separate
zone_cut_root and zone_cut_deleg flags, as this makes some logic
much clearer, vs interpreting the meaning of "zone_cut" depending
on the tree context as we walk nodes.  It's also imperative that we
don't need that context to decide which subclass's structure was
written to, which the zone_cut_root flag is doing for us.