Fix log output for binary bytes in domainnames
The dname-to-string conversion function used to write domainnames
for log message outputs had a bug which caused it to incorrectly
print the DNS-style decimal escapes for binary data bytes in
labels.  The cause is that a data type was 'char' when it should
be 'uint8_t' for the math that creates the escapes to work
correctly.  The bug was introduced back in ~2014!

(cherry picked from commit 29973d97c42ee8989d52621903e5bc1172488969)