Replace volatile sig_atomic_t with C11 atomics
This gets rid of the last remaining use-case for "volatile" in our
code, which was flags set by signal handlers.  These can be
replaced with relaxed access to atomic_uintptr_t.

Any type with at least one bit would do, and atomic_flag is
gauranteed to work, but its API is clunky with no way to test
without re-setting and a backwards (for our purposes) initial
value, and we're already requiring lock-free uintptr_t for other
purposes anyways, so using that is simpler for us.

[Also, reworded a sentence in the docs to get rid of the word
"volatile", so that "git grep volatile" comes back empty for the
whole repo now]