Clean up socket handoff error reporting
All error paths here are ultimately fatal, so we can upgrade all of
these "log_err(...); ...; return 0" to immediate log_fatal at the site
of the problem, clearing up the logic a bit and ensuring all of the
possible fatal messages here are consistent and unique.  A couple more
minor new constraints are added to avoid letting corner cases slip
through as well.