Add --with-buildinfo=X to override "Build Info"
This gives packagers/builders the ability to override the "Build
Info" string which is output with various CLI help messages.

"Build Info" usually contains either a `git describe` -derived
version number, or the fixed string "non-git" if no such info can
be found.  This is problematic for some distro build systems that
sometimes build from non-git sources, but need the "Build Info" to
be consistent so that the binaries (which contain it as a string)
are consistent.  They can use this option to supply their own
version string, perhaps a vendor-specific version string derived
from the package's version control.