Refactor a loop var from signed to unsigned
Some modern compilers, with some combinations of flags, would emit
a warning about a possible signed overflow of "j" in the original
version of this loop, during some optimization pass.  It's not
hard for a human to prove to themselves, based on asserted
constraints and basic logic, that no real overflow bug happens
here, but it's concerning that the compiler is confused and thus
might mis-optimize, so it's probably better to just refactor it as
unsigned.

In this version, the now-unsigned j can wrap from 0 => UINT_MAX in
the final decrement of the while loop (unless we terminate the
loop earlier based on the sort condition), which is then caught by
the "j < i" condition at the top if that happens, and then the
final use of j adds one back to it before using it as an array
index, making it zero, so everything's as fine in this version as
it was before, but the compiler doesn't complain.