UDPv6: always clear ipi_ifindex for non-LL addrs
This was originally noted and fixed just for FreeBSD in #174 a
couple of years ago with a patch from Ryan Steinmetz.  On further
analysis, I think his patch is actually the more correct thing to
do for all platforms in the general case, not just FreeBSD.
Without it, the same case the original reporter ran into (problems
with global unicast addresses bound to the loopback and routed
into the host explicitly) probably happens on at least some other
platforms, and may happen with others in the future.

This also invites questioning whether we should even be using
IPV6_PKTINFO cmsg data for all IPv6 listen sockets as we do today.
You could make a logical argument that it's only necessary at all
when the bound address is either ANY (as is this case for IPv4) or
is explicitly a link-local address.  I'm not 100% sure of that
conclusion though, and continuing to do it for all IPv6 sockets
like we do today is at worst just a very minor inefficiency, I
think!

(cherry picked from commit 45191aa9f8f64be1874cdb08cde78ec154d38af8)