UDPv4: Avoid setting up cmsg for non-ANY sockets
For UPDv4, only sockets bound to ANY actually need any cmsg data,
to carry either of IP_PKTINFO or IP_RECVDSTADDR depending on the
platform.  What we've done so far out of convenience is to set up
the cmsg buffer space fully for all sockets, meaning that before
our recvmsg() call, for all cases msg_hdr.msg_control points at a
real buffer and msg_hdr.msg_controllen is set accordingly.  For
the non-ANY case, we simply didn't set the appropriate sockopt to
ask for any data there and didn't really expect any to arrive.

We also re-use the same msg_hdr structs from recvmsg() as
arguments to sendmsg(), because this is the simplest way to pass
off the source address information from recvmsg() when necessary.

In the non-ANY case, this usually means that recvmsg() will set
msg_controllen to zero (because it didn't supply any control
messages, because we didn't set the sockopt to ask for any), and
then we end up invoking sendmsg() with the same msg_controllen of
zero, but with msg_control still pointing at our
allocated-but-empty buffer.

However, somewhere between FreeBSD 12.0 (the last I had tested on)
and FreeBSD 12.2, FreeBSD has become more strict about the inputs
that sendmsg() finds acceptable, and so in these non-ANY cases,
sendmsg() ends up failing with EINVAL because it doesn't like us
sending over a non-NULL msg_control buffer pointer alongside the
zero msg_controllen value that was given to us by recvmsg().

The fix here (which is a bit deja-vu, because we used to have a
similar setup for different reasons in the distant past) is to
create a "use_cmsg" argument to the mainloop functions which we
set false for IPv4 non-ANY sockets, which then suppresses ever
setting the useless msg_control buffer pointer in the first place
for the IPv4 non-ANY case.

This seems to make things work for me in a FreeBSD 12.2 VM, and
doesn't seem to break Linux, either (not that we'd expect it to).

Fixes #202 (thanks for the bug report cstdenis!)

(cherry picked from commit f2c6a65744b5567ba55f939d6e103be7196238a4)