x32: use correct format specifier for stats JSON
The format specifier for JSON stats output was fixed as PRIuPTR,
but we need it to be PRIu64 in our special case for the x86_64 x32
ABI.  This only affects x32 builds, which I don't think is very
popular, and in any case apparently modern compilers would DTRT
anyways after warning about it, because stats-dependent tests
succeed on that platform.  Probably x32 stats were effectively mod
2^32 in the output prior to this commit, but they'd still be
correct in all other senses.