Warn about default listen=>any in docs and logs
We may need to change our default listen=>any behavior at some
point in the future, as it's not ideal for OS/distro packages
dealing with the intersection of possibly several unrelated DNS
daemons being installed (e.g. a local stub resolver, an
authserver, and some kind of cache all trying to bind various port
53 by-default).

I haven't found a great way to change this in 3.x without either
breaking useful setups users would care about or adding way too
much hackery and complexity just to limit the impacts imperfectly.

For now, we'll just warn about this in the docs and the log
outputs, in hopes of prodding users to stop relying on the default
behavior ahead of probably breaking it in a future major-version
upgrade.