Enhance some daemon-replace error messages
These message are all relevant to debugging any "replace" issues,
and some of them were lacking in detail and/or lacking the
standard "REPLACE[old daemon]:" prefix to make the log flow easier
to read.

In the particular case of an illegal TAK2, also added a close of
the offending client connection.  This does not affect correctness
on either side, but is more consistent and there's no good reason
not to close here.