TCP: quiesce RCU between responses
This adds an additional RCU quiescence point in the TCP I/O
threads once per response.  The overhead is trivial (esp for TCP)
and it may help make things smoother in some edge cases where a
single eventloop iteration is doing an extraordinary amount of
work for many connections with a high tcp_clients_per_thread.