TCP: stricter registry management
ThreadSanitizer pointed at a potential data race between
dnsio_tcp_request_threads_stop()'s access to thr->stop_watcher and
the free(thr) that happens just before a tcp thread actually
exits, and this commit "fixes" it by adding an unregister function
and putting explicit locks around all related operations.

Technically, I don't think there was a real bug here that anyone
could hit at runtime.  There are logical constraints the sanitizer
can't see: that the free() can only be reached during clean
termination, which is triggered by request_threads_stop anyways,
thus ensuring that the stop function's access to thr happens
before (causes) the eventual free(), combined with the fact that
the stop function is only called once from main().

But still: it's a bad smell and I'd rather just be over-cautious
and future-proof, since adding these locks and calls doesn't
harm anything in the perf-critical fast paths.

We already fixed the other side of this once (the sequencing at
startup) in f5ea429ca a couple years ago, so it's not like there
isn't potential for further bugs, especially if we make more TCP
changes down the line and get confused.

[Side note - in general the current code isn't amenable to
analysis with ThreadSanitizer, because TSan doesn't grok what
we're doing with various lock-free mechanisms like our
stats-gathering hacks and our liburcu data swaps + reclaims.
However, I do have some prototype code which uses C11 atomics for
the stats and swaps out userspace-rcu with a C11 atomics variant
that's just for this kind of testing as well, which cleans things
up well enough that TSan is no longer confused, so that I can use
it to look for issues like this one.  I think those bits will make
it into 4.x, so that we can easily build a development copy of the
code that's TSan friendly, but for now it's not easy to try TSan
on the master branch for yourself because of all the RCU/stats
false-positives]