More efficient idling for UDP threads
Idle UDP threads now block indefinitely in a ppoll() syscall if
there's been no traffic for ~257ms, and in the worst case this is
also the maximum possible (but unlikely) delay we'll impose on RCU
writer reclaimations and shutdown requests.

Prior to this commit, all idle UDP threads would wake up every
~3.1 seconds to re-execute recvmsg(), and in the worst (but very
unlikely) corner case could delay daemon shutdown by up to that
same ~3.1 second duration, but the max RCU delay was ~109ms.

This should be better a better set of tradeoffs all around, and
especially helps with efficient power/resource management in the
case of unused UDP threads (e.g. unused listen addresses, or
unused daemons in general), but it's a significant restructuring
of the main UDP loops, so the change is not without some new risks
of new bugs!