Prevent rare double-free/crash in late shutdown
On daemon shutdown (due to stop or replace), if a reload-zones
operation was still ongoing at the time and a rare inter-thread
timing race came out just wrong enough, it was possible for the
reload-zones thread and the main thread to both attempt to destroy
and free the same final copy of the zone data, leading to some
kind of double-free and/or crash as the daemon was trying to exit.

This happens very late in the process shutdown game after all DNS
listeners are stopped, so there's no 'real' effect for clients.
Still, it's better that we avoid the possibility of the daemon
crashing out at the end like this.

Because deallocation was happening before stats handoff, the crash
would also prevent handing off stats to the new daemon for
continuinity in the replace case, but regardless the replace
operation would still ultimately succeed with a stats rollover.

I first tried to fix this by explicitly synchronizing a join of
the final reloader thread with the ltree cleanup, but I later
realized that a far simpler and more-efficient answer was to just
eliminate the ltree_cleanup function.  Its only purpose was to
exercise tree deallocation for test or QA purposes, as either way
this is global-scoped unleaked allocations that will go away with
the process's death, and we have better ways to test these things
already in place.

With this pointless deallocation exercise removed, the whole race
goes away and we don't have to care about the state of any running
reloader thread at shutdown time.  A similar (but at least not
racy) argument applies to why we're bothering to clean up
challenge sets, so that atexit deallocation function was removed
as well.