Replace unaligned macros with memcpy wrappers
This gets rid of our last reliance on the non-standard "packed"
attribute (which really isn't guaranteed to work the way we expect
here for all platforms and compilers anyways, even when it does
exist!), by replacing our unaligned access macros with small
inlinable functions that wrap memcpy().

This is more portable and reliable, and most optimizing compilers
for most platforms generate the roughly (or exactly!) the same asm
code for this as they did for the previous solution.  In the rare
case this gives less than optimal results, it's up to the compiler
to do a better job!