Define numeric package version in autoconf
This gets rid of the pointless use of a runtime sscanf() to parse
our version string (which is fixed at build time) into numeric
values, by doing this during the configure run.  Along the way, we
needed to parse it correctly in the configure script, and thus
we've added some much-stricter validation on the PACKAGE_VERSION
string as well (that the canonical part conforms to semver rules,
and that the major, minor, and patch numbers are in a reasonable
range, etc).

None of this affects --with-buildinfo or the autogenerated build
info output.  For better or worse, the "build info" string in the
daemon help output is completely independent of the autoconf-level
official PACKAGE_VERSION.  It seems like it would be nice to bring
them together under a single source of truth somehow, but there
are lots of edge cases that make it seem not worth it, and the
current scheme "works" for all the known, common, good-faith
scenarios.