Document the control socket protocol
This might help third parties wanting to write control socket
clients for operational reasons.

Having to write all of this down in human terms also helps us with
future development, as it can be hard to piece all this together
just from reading the source code, especially for the inter-daemon
takeover sequence!

The first draft of such a long document is never perfect, but this
is a good starting point anyways!