Internal renames of some terminology
I'm aware that in tech circles there's some controversy about how
these sorts of renamings towards less controversial terminology
can be controversial in themselves.  This is especially true in
the cases where nobody was meaningfully complaining in the first
place, and when the change has the potential to cause a bunch of
breakage, churn, and/or work for a lot of other people.

However, the changes here are entirely internal to our codebase
and don't break workflows for anyone else, and more importantly
the spirit behind these changes is that if I were writing these
lines of code today, I could and would have been more
conscientious about using different terminology choices in the
first place.  These changes also don't, in my opinion, reduce
clarity or misalign us with any upstream/external namings in e.g.
RFCs or libraries, etc.  In some cases (the SOA record fields)
they actually increase alignment with the RFCs in this case.

In any case, this change is entirely a mechnical and should not
have any impact on correctness.