Fix some testsuite races
There are a handful of tests related to statefiles which implement
a common pattern of writing a statefile, monitoring the daemon's
log output to see that the file was (re-)loaded, then executing
test DNS queries to witness visibile DNS-level effects.

However, in many such cases, the monitored log output did not
technically guarantee the state change effects are visible in DNS
outputs yet.  These tests commonly succeeded anyways, but they
could very rarely fail because of a testsuite timing race.

In all but one such case, this was trivially fixable by only
changing the testsuite to wait for an additional log output line
which actually does guarantee visibility in the tested case.

The tricky one was the case where the "admin_state" file
previously existed and had meaningful contents, and then was
deleted from the filesystem.  In this case, the daemon did not
output any log line that could be used for testsuite
synchronization in this manner, so we also had to add new log
outputs to the daemon itself to facilitate reliable testing.

Many thanks to debian's QA infrastructure for finding a rare race
causing an FTBFS during an attempted reproducible build, which
lead to me tracking this down!