Minor optimizations for iothread startup
This is perf nitpicking of little consequence, but technically
this is a little more efficient at getting through the
condvar-gated part of iothread startup.  The main thing is moving
the xcalloc call out of the mutex-protected critical section, thus
shortening the time the mutex is held.  The rest is just slight
re-organizations of relatively-unimportant orderings.