UDP: do not block on send
This changes the send[m]msg() calls in the main UDP I/O loops to
use MSG_DONTWAIT, and changes our error handling for these calls
to not retry on EAGAIN or EWOULDBLOCK.

Both approaches are valid and everything's the same under "normal"
conditions, but they have different tradeoffs under extreme
traffic pressure where the NIC can't drain the UDP socket's output
buffer fast enough.

Neither strategy can completely prevent the effective loss of some
responses when buffers and/or NIC speeds are insufficient for the
traffic.  Blocking on the output side favors conserving CPU
resources at the cost of potentially delaying responses to the
next several packets (and if the input buffer fills up, then
losing some of the newest rather than the oldest input packets),
whereas a non-blocking output side "wastes" CPU generating
responses that are dropped when they can't be sent in order to
avoid introducing unecessary delays in processing the next
request(s).

This seems like the better set of tradeoffs to me at this point,
and more-closely matches how we handle things in the TCP case as
well.  It does mean that excessive input packets will costs more
CPU in practice, but it also ensures we don't unecessarily delay
later requests in case of a temporary spike.