CDL: generic counted doubly-linked list macros
The primary initial use-case for this is the connection timeout
queue management in the TCP code.  That has been using a hacky
inline custom doubly-linked-list implementation all this time, and
it would probably be cleaner and simpler to use a generic set of
macros like this.  We might find other use-cases in the future
and/or need to expand the functionality offered.

This CDL structure also includes an explicit counter for the
length of the list, as the initial case and at least a couple of
others will want this anyways.

We may end up defining some simpler variants for some cases (e.g.
single-linked and/or no-tail-pointer and/or no-counting), but
first we'll transition all reasonable cases to this structure and
see what the fallout looks like for the rest of the code, and then
perhaps define and transition to lighter variants as warranted by
any efficiency concerns.