Add a CACHE_ALIGN macro for use with alignas()
There are probably a few (at least one I can think of!) cases
where we should be cache-aligning some structs that are accessed
x-thread to avoid false-sharing.  This adds some autoconf stuff to
define a CACHE_ALIGN macro the rest of the code can use.  Platform
detection is pretty minimal guesswork at this point and might need
further iteration, but there's a configure-time override builders
can use and a reasonable default of 64 for platforms we don't have
custom defines for.