TCP: Use CDL for connq, make timers more efficient
This switches out all the hacky custom doubly-linked list code
used for TCP connection queue management for the new CDL_*
macro-based system.

Timeout timer management was also made more consistent and
efficient along the way by making it lazier.  The timer is only
re-set (or stopped) when strictly necessary: right after it fires,
and when adding the first new connection while it's stopped.  For
the rest of the normal runtime execution, it's better to leave it
running and let it potentially fire "early" than to constantly
churn its value while various connections are rapidly created,
destroyed, and/or activity-bumped.