Break up the dnspacket_wait_stats gate
TL;DR - Everything here is inter-related around the core idea of
making thread and global startup sequencing more asynchronous and
efficient by getting rid of a pthread condvar gating system that
was synchronizing all threads' execution reaching a certain point
before the main thread could finish starting up.

Background: ---

  It all started with how stats were handled.  Historically we
pre-allocated an array-of-pointers (one per i/o thread) and then
had each thread allocate its own stats structure and register it
into the shared array that the stats reader code consumed from.
To avoid races and breakage, dnspacket_wait_stats() was created,
which was a pthread condvar-gated system of ensuring all i/o
threads had completed their (pthread-locked) registration into
this array before letting the global startup sequence move forward
to where the stats output code could attempt to iterate the array,
thus ensuring safety.  Some of the safety needs here were relaxed
over time, but the mechanism remained because it was reused for
other things:

  At the TCP layer, we also developed a need to create a thread
registry to handle the clean sequencing of TCP thread shutdowns at
daemon stop time.  This was also done as a pre-allocated array,
and when it was clear that this had similar race issues to the
early stats situation (what if the main thread attempts shutdown
before a TCP thread has finished starting up and registering with
the shutdown-handling array?), we reused the dnspacket_wait_stats
gating system from above to also gate these safety concerns at the
TCP layer (and then later tossed an extra TCP-layer mutex on top
just for reassurance).

  Another thing the dnspacket_wait_stats gating system was doing
was ensuring that all DNS sockets were "listening" for requests
before we informed the system and the user that the daemon was
"up" and they could rely on it answering requests.  For UDP
sockets we didn't really need the gate (they are queueing requests
from the moment they bind, which happens earlier in single
threaded code), but for TCP the "listen()" call was happening
early in a thread's asynchronous startup, and so the gating system
was also assuring us that this had already happened.

The Key Changes: ---

* The stats registry and TCP thread registry had their array of
  pointers systems replaced by simple linked list designs.  In
  both cases, each thread adds itself to the list once at daemon
  start, and nothing is ever deleted from the list.
* Because the stats output code can be successful and safe even if
  it executes before all threads register their structs to the
  linked list, it no longer needs the condvar gating.  This case
  is so simple that even the linked list's mutex was replaced with
  a trivial lock-free CAS system.
* The TCP listen() call was moved up to single-threaded socket
  setup code, thus removing the need for the gating system to
  ensure that TCP was accepting requests.
* The actual mutex+condvar gating system was now only necessary
  for the TCP thread shutdown case, so the condvar-gating part was
  moved to the TCP shutdown function itself to ensure safety if
  shutdown happens very early while threads are still starting up,
  and thus the gating no longer happens in the startup sequence of
  the daemon at all, and commonly for longer-lived daemons it
  won't stall at stop time, either.

The net of all this is that, without losing any real safety, there
is now no critical point in the main thread startup sequence that
has to condvar gate on all the I/O threads reaching a certain
point in their own executions, and so the whole daemon's startup
can proceed in a more asynchronous and efficient way.

Some other minor related bits were re-arranged in the process, and
the per-thread stats structures were given some cache-alignment
magic to assure that unrelated writes to thread-local memory
weren't causing false sharing with the main thread's read of the
per-thread stats structures.