UDP: simplify send-side code
This isn't intended to cause any functional change, but it makes
the flow of control more human-readable (IMHO) and possibly more
optimizable as well, especially in the sendmmsg case.

The previous version of the sendmmsg loop did a fair amount of
redundant tracking (e.g. "foo--; bar++;" sorts of patterns, and
the redundant usage of both an iteration pointer ("dgptr" for
"dgrams") and an index integer on the iteration pointer), while I
think this version distills the essence of the logic more-clearly.