Coding style: remove most of the typedefs
This is another case of my aesthetics about C code changing over
the years, and it being finally time to bring this codebase up to
date with them during a major release cycle.

I used to really like the "typedef struct { ... } foo_t" idiom.
However, Over the years I've come to think of it as a bad pattern,
but I've continued using it in this codebase up to this point in
the interest of self-consistency.

This commit rips out the vast majority of such typedefs for enums,
structs, and unions from the code, setting us up to move forward
with better style.  The few typedefs that remain are used
sparingly for cases such as:

* Function pointer prototype typedefs for callbacks (because
  they're so ugly and error-prone to use without a typedef).
* Some integer typedefs remain for special-case purposes that
  generally have accessors and important semantics (atomics, funny
  bit usage)
* In one remaining case, I've left a typedef in for a really
  strong abstraction barrier over a union type (vscf_data_t)

For all other cases, I don't think the typedefs are a net win,
especially as they hide important information about the nature of
the type (that it's a struct, union, or enum) which can aid human
analysis.