vscf: better error output for bad $include
When an include statement appears in value context, it cannot be a
glob or directory name which includes more than one file.  The
check for this was faulty and allowed for two matches, but also it
was dead code because the condition was being caught in a
different part of the logic and messaged inappropriately.  This
should fix up the logic to say the right thing at the right time.
Either way, all such cases ended up being parse failures.