vscf: handle DNS unescaping errors more-gracefully
For the vscf config parser, the code which unescaped any DNS
escaping in "simple" values can return zero for a failure to parse
the DNS escapes correctly.  The scanner was effectively ignoring
this and causing the creation of a zero-length (just a NUL) string
in this case, which would cause an error to appear elsewhere in
the consuming code. Now it properly throws a parse error at parse
time instead.

Fixing this without messing with a lot of interface assumptions
meant moving the unescaping to happen at parse time.  Previously
unescaping happened at the time the value was used (because
there's one usage function vscf_simple_get_as_dname() which
applies a separate and stricter parsing).  The dname case now gets
parsed twice, but this is a minor inefficiency.

As a result of the early-parsing decision, the cloning function
for simple values was also re-written so that we don't pointlessly
re-parse in that case as well.