scanner bugfix for large zonefiles
In 64-bit builds, a bug was preventing the correct parsing of
zonefiles with more than 4GiB of data in them.  In such cases, the
bug would cause the parser to only parse the first "size mod 2^32"
bytes of the zonefile, which is likely to be unsuccessful, but
could also "successfully" load partial data if the modded size
landed at a legal point between records.  Basically, an "unsigned"
was being used where a "size_t" was intended in a function
signature.