Remove valgrind suppressions for replace
Switching out an _exit() for an execl("/bin/true") when running
under valgrind gets rid of the need to suppress leaks related to
the DNS I/O threads' allocations due to the double-fork pattern
that happens during "replace".

Also, while in here, increased the specificity of the remaining
suppression to ensure it's not hiding unintended cases.