Upgrade some log_devdebug to log_debug
These are not in perf-critical areas and not driven by client
network traffic, so categorically they belong better in the
log_debug class than the log_devdebug class.  This leaves the only
cases of log_devdebug in the dnspacket code, which aligns better
with Future work on better logging controls.