Robustness improvements for replacement launching
There were several unlikely corner case situations where the
daemon replacement sequence was not as robust as it could be.

None of these have caused an observed failure in the wild, to my
knowledge, but it is possible.  Mostly these are cases that would
only happen on a faulty or resource-exhausted system (e.g.
pipe2() failure, fork() failures, pthread_sigmask() failures,
etc), but there was also a case where something as simple as a
missing replacement binary could still result in the fatal exit of
the original daemon.

The cases fixed here fall broadly into two categories:

* There were some unsafe error logging calls during the
  replacement fork()->exec() window, which happens while other
  threads are running and thus must only call signal-safe
  functions.  The unsafe logging calls could cause undefined
  behavior in the original daemon.  These unlikely error outputs
  have been replaced with safe writes to stderr with fixed error
  strings.
* There were some cases where the new replacement daemon would
  fail to launch for an unlikely reason and the result would be a
  [safe] fatal error in the main daemon, leaving no daemon running
  at all.  These have been fixed to merely log the errors and
  continue the operations of the old daemon.

Along the way, some minor structural changes were made for
clarity, the libev default loop's automatic SIGCHLD repear was
removed, SIGCHLD was set to SIG_IGN for the main process, and the
process-killing code (which wasn't actually useful in the previous
design) was fixed to work correctly now that we're actually
relying on it.

The log messages after failing to launch a replacement are now
more-informative as well, in the sense that they'll indicate
whether there was a definite failure-to-launch vs a replacement
process that may or may not be running correctly, but which we've
merely lost the ability to track due to some meta-failure.