Move dclists struct def to header
CodeQL has been complaining about a build failure here, as I
suppose it needs more type information that basic compilation
does.  This type was meant to be an opaque pointer outside of
dclists.c, but it's not terribly important that it remain that
way.  I'm not even sure if this is the only such example, so I may
need to do several similar followups as it uncovers them on runs.