tests: fix OPT-related Net::DNS 1.36 breakage
With Net::DNS 1.36, code such as:
  $optrr->option('TCP-KEEPALIVE' => pack('n', 370));
is now broken, with the function doing the packing internally, and
therefore expecting something like:
  $optrr->option('TCP-KEEPALIVE' => 370);

It gets even murkier for more complex structures like CLIENT-SUBNET. Plus, it's
not clear how we could construct invalid binary data, like we are to generate
e.g. too short options.

This is likely an upstream regression that may be fixed in the future. However,
1.36 has now found its way into downstream distributions, cementing the issue
for years to come.

Thankfully, working around this is not too bad: one can mess directly with the
internals of Net::DNS::RR:OPT, and set the option in the hash directly. As this
is an internal data structure, it may break in the future; it's not like
published APIs are stable, though, so we can cross that bridge when we comes to
it. It is, at least, backwards-compatible: tested with 1.05 and 1.32 - 1.36.

Introduce a new _GDT::optrr_option_set helper function for ease of use, and
convert all users.